+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - New model: `PhysicssrModel`
 - New model: `DrnModel`
+- New model: `HanModel`
 
 ## [0.1.4] - 2021-08-17
 ### Added